+++ tests/test_at_src_001_simple.py
@@ -38,14 +38,14 @@ def test_sourcefile_parsing():
 
         # Check directions are normalized
         norms = torch.linalg.norm(directions, dim=1)
-        torch.testing.assert_close(norms, torch.ones(2, dtype=torch.float64))
+        torch.testing.assert_close(norms, torch.ones(2, dtype=torch.float32))
 
         # Check wavelengths
         assert wavelengths[0].item() == pytest.approx(1.0e-10)
         assert wavelengths[1].item() == pytest.approx(1.5e-10)
 
         # Check weights are preserved from file per AT-SRC-001 requirement
-        expected_weights = torch.tensor([2.0, 3.0], dtype=torch.float64)
+        expected_weights = torch.tensor([2.0, 3.0], dtype=torch.float32)
         torch.testing.assert_close(weights, expected_weights)
 
         print(f"✓ Source file parsing successful")
